:truck: Move `solidity_version.py` to `core` module
