+++ tests/test_config.py
@@ -1,10 +1,10 @@
-from pathlib import Path, PosixPath, WindowsPath
+from pathlib import Path
 import pydantic
 import pytest
 
 from woke.config import WokeConfig
 from woke.config.data_model import SolcRemapping
-from woke.regex_parsing.solidity_version import SolidityVersion
+from woke.core.solidity_version import SolidityVersion
 from woke.core.enums import EvmVersionEnum
 
 sources_path = (Path(__file__).parent / "config_sources").resolve()
